Initial commit for dissect.executable.pe
First version for the dissect PE format parser. Currently the parser
supports standard PE files, as well as some patching of executables and
building an executable from scratch. It is expected there will be some
issues that will be found during usage and it will need quite some
refactoring. The API as it currently exists might need some revising too
:)